+++ services/backend/src/main/java/com/soprasteria/g4it/backend/apiadministrator/controller/AdministratorRestController.java
@@ -86,7 +86,7 @@ public class AdministratorRestController implements AdministratorApiDelegate {
      */
     @Override
     public ResponseEntity<OrganizationRest> createOrganization(final OrganizationUpsertRest organizationUpsertRest) {
-        return new ResponseEntity<>(organizationRestMapper.toDto(administratorOrganizationService.createOrganization(organizationUpsertRest, authService.getAdminUser())),
+        return new ResponseEntity<>(organizationRestMapper.toDto(administratorOrganizationService.createOrganization(organizationUpsertRest, authService.getAdminUser(), true)),
                 HttpStatus.OK);
     }
 
@@ -123,7 +123,7 @@ public class AdministratorRestController implements AdministratorApiDelegate {
     @Override
     public ResponseEntity<List<UserInfoRest>> linkUserToOrg(final LinkUserRoleRest linkUserRoleRest) {
         return new ResponseEntity<>
-                (userRestMapper.toListRest(administratorOrganizationService.linkUserToOrg(linkUserRoleRest, authService.getUser())), HttpStatus.OK);
+                (userRestMapper.toListRest(administratorOrganizationService.linkUserToOrg(linkUserRoleRest, authService.getUser(), true)), HttpStatus.OK);
     }
 
     /**
@@ -132,7 +132,7 @@ public class AdministratorRestController implements AdministratorApiDelegate {
     @Override
     public ResponseEntity<List<UserInfoRest>> updateRoleAccess(final LinkUserRoleRest linkUserRoleRest) {
         return new ResponseEntity<>(
-                userRestMapper.toListRest(administratorOrganizationService.linkUserToOrg(linkUserRoleRest, authService.getUser())), HttpStatus.OK);
+                userRestMapper.toListRest(administratorOrganizationService.linkUserToOrg(linkUserRoleRest, authService.getUser(), true)), HttpStatus.OK);
     }
 
     /**
@@ -153,5 +153,6 @@ public class AdministratorRestController implements AdministratorApiDelegate {
         administratorOrganizationService.deleteUserOrgLink(linkUserRoleRest, authService.getUser());
         return ResponseEntity.noContent().<Void>build();
     }
+
 }