+++ services/frontend/src/app/core/interfaces/administration.interfaces.ts
@@ -55,3 +55,15 @@ export interface OrganizationCriteriaRest {
     criteriaIs: string[];
     criteriaDs: string[];
 }
+
+export interface DomainSubscribers {
+    id: number;
+    name: string;
+}
+
+export interface Workspace {
+    id: number;
+    name: string;
+    status: string;
+    defaultFlag: boolean;
+}